scripts: add script for improving find-replace
This is a helper script for finding and replacing the unique hashes in
our tests, primarily the `radicle-cli` tests.

Whenever we change the format of something that produces a hash, many
tests can end up failing since the hashes have changes. This script
allows you to specify a list of replacements in one go to more quickly
fix the tests.