cob: use underlying git_ext::Oid for serde
The use of bytes for serializing and deserializing was incorrect. It means that
the serializing and deserializing of `ObjectId` and the underyling
`git_ext::Oid`, which tends to cause issues with consumers of the API.

Change the `Serialize` implementation to defer to the `git_ext::Oid`
implementations.

For `Deserialize`, an internal enum is used to deserialize into a `Vec` or
`str`, which can then be converted into the `ObjectId`. This is to ensure
backwards-compatibility with `ObjectId`s that have been serialized into the
bytes form already, e.g. the COBs cache.

Add `qcheck` tests to ensure that these two types are compatible.