cob: introduce merge entries
The purpose of this change is to allow for an API where a remote can perform
draft changes to a collaborative object in parallel to their published changes.
When the changes are ready, they can merge their two tips to form a new entry,
which is used to correctly build the change graph. When evaluating the history,
only the `Entry` changes are considered, since they contain the important
`contents` information.

To support this, the APIs needed to change to account for loading a
`MergeEntry`. Thus, the new return type for `change::Store::load` is a
`ChangeEntry`, which is an enum of `Entry` and `MergeEntry`.

In turn, the internals have changed to use `ChangeEntry` in most places as
opposed to `Entry`.