radicle: diff configuration data
Introduce diff configuration data so that it can be used reproduce diff
information in a stable way.

This information can be stored alongside a patch so that when comments are made
on a patch's diff, then their positions can be stable across clients – as long
as they reproduce the diff with the same configuration.