cob: introduce Filter for filtering entries
The aim of this change is to introduce filtering of entries by taking advantage
of the `prune_by` method during the evaluation via `ChangeGraph::evaluate`.

This is the minimal change to introduce a `Filter` struct that can be passed
into the `evaluate` method. It has a `Default` instance, and this is used to
introduce the change without any change to the semantics to start off with.