radicle: refactor rule matching
This change improves the API for rule matching by providing a method that
produces a `Canonical` directly from the set of rules – as opposed to going
through the previous `MatchedRule` struct. `Canonical` now carries its
`ValidRule`, which means that invalid states regarding threshold and the
allowed set of DIDs are unrepresentable.

It also provides a method `matches` that returns all matching `Pattern`s. This
allows us to provide the CLI subcommand, `rad cref match` that shows which
rules match for a given reference.