test: set name and email for repository fixture
In fresh environments, without a global git config, the `user.name` and
`user.email` will not be set for the fixture repository. This causes issues when
tests are run and attempt to call `git2::Signature::now` – it will fail to read
a user name and email for the commit.

To remedy this, the name and email are set for the repository when initialising
it.