radicle/patch: Add Change ID
Radicle patches have IDs, which are the commit IDs of the COB.

Other tools, like Jujutusu, Gerrit, and GitButler use Change IDs to
refer to commits consistently as they are rebased.

In order for Radicle to integrate with these tools more easily in the
future, add an action that allows linking a patch with a Change ID.

This naturally requires an optional struct member and some boilerplate.

See also:
https://lore.kernel.org/git/CAESOdVAspxUJKGAA58i0tvks4ZOfoGf1Aa5gPr0FXzdcywqUUw@mail.gmail.com/T