protocol: session state
This change introduces session state as part of the sans-IO rewrite of
connection management. It is similar to the `Session` and `Sessions`
types that are defined as part of `service`.

Where the `Session` and `Sessions` type differ is by providing a
generic type for keeping track of which state they are in: initial,
attempted, connected, or disconnected. The nodes in each state are
then kept track of in individual maps – one for each state.

The transition logic and mutation of session state are all
encapsulated in this module, so as not to leak these details.

The other benefit of using the generic parameter allows us to define
methods that are specific to a specific state for a session,
e.g. being able to ping a connected session.