protocol: peer connections
An event/command approach to keeping track of peer connections.

The `session` module keeps track of the state of different peer sessions in
several maps. It keeps them separated by the state that they're in and ensures
that the correct type of session is used using a generic type.

These sessions are then kept track of in a `Connections` state machine, that
performs the checks of what should happen in the state machine, `commands`, and
output useful `events`, and `effects` that can be observed by the rest of the
protocol.