protocol: connections manager
This change defines the connections manager in a sans-IO fashion.

The goal is to have on-par behaviour with the management of connection
sessions that exists in the current `Service`.

It does this by using the `Sessions` state, defined in the previous
change, and pairs with commands and events.

The commands indicate what inputs the `Connections` state can receive
for changing the state of the sessions, and the events are the outputs
for those changes.