protocol/service: wire up connections management
This replaces the `Sessions` logic by using `Connections`.
Wherever there was a use of `Sessions`, `Connections` is now used to
drive the state forward.

There are two ways we use the commands, paired with the I/O actions.
The first is to provide the command before the I/O action, so that the
state can be checked and validated that it can transition,
e.g. `connect` to see if a connection can be made.
The second is the I/O action drives the updating of the state that is
reflective of the real-world, e.g. `connected` since an inbound
connection was made. Note that further actions may be take from events
output here, for example, disconnecting the connected node because the
running node is at its limit.

It also removes the configuration constants in `service` to
consolidate them into the one module `connections::config`, and reuses
values set via `connections.config()`.