REVIEW(WIP): refactor changes
Capture the `BTreeMap<Oid, u8>` type into a `Votes` struct, with its own API.

We were getting the objects from the repository twice – once in `Canonical::new`
and once again in `ensure_commit_or_tag` – so the latter was removed.

Use an `enum CanonicalObject` to specify which object types the canonical
process supports.

During `converges`, separate commits and tags, and ensure that we are only
looking at objects of one type. Note, that I think this is actually the *wrong*
place to do it because we already filter the objects – so we should keep track
of that information higher up the callstack.

WIP: attempting to fix the tests.