protocol: fetch state
Similar to previous efforts, this change takes the fetching slice of the
`Service` and attempts to capture a sans-IO implementation of it.

It essentially translates `Service::fetch` into the `FetchState` machine. What's
important to note here is that it does not entangle the idea of sessions with
the idea of fetching. The state is purely based on `NodeId`s and `RepoId`s.

Note here that a layer above would make sure that connected sessions, using the
`Connections` state, would perform fetches and using `FetchState`.