protocol: node events
The protocol will often want to emit node events so that a subscriber
can listen to these events and act upon them – for example, the
Radicle CI code uses this to decide when to start CI jobs.

The sans I/O approach here is a simple queue where events can be
pushed, popped, or drained.

An `Emitter` trait is provisionally defined to show that these events
should be emitted from a producer to some subscriber or subscribers.