protocol: connections subscribe commands
Add commands for mutating the subscription of a peer session.

One is intended for setting the subscription, expected to be used when
a subscription message is announced as a gossip message.

The other is inserting an RID into its filter set, expected to be used
when a gossip message from the node gives us information that this
node is interested in the given RID, e.g. refs announcements,
inventory announcements.