protocol: routing handler
The method `Service::sync_routing` handles setting an announced
inventory from a node and keeps track of whether the RIDs are added,
updated, or removed. It does this by interacting with the routing
table.

To translate this into a sans I/O approach, the inventory is tracked
as a BTreeSet, that can then be compared against using
`BTreeSet::difference` calculations.

The effects of setting the inventory in the routing table is then
captured by a trait.