protocol: fetch announced
Encapsulating the behaviour of checking an announcement to see if any of the
signed references are what the local node wants.

In the previous `Service` implementation, this was called
`Service::refs_status_of`.

Since the original logic required accessing sqlite, the business logic is
captured within a state machine, that can then be driven by traits the capture
the calls to the sqlite db.